# cmd/api/users.go
package main

import (
	"context"
	"errors"
	"gopher_social/internal/store"
	"net/http"
	"strconv"

	"github.com/go-chi/chi/v5"
)

type userKey string

var userCtx userKey

// @Summary		Fetches a user profile
// @Description	Fetches a user profile by ID
// @Tags			users
// @Accept			json
// @Produce		json
// @Param			id	path		int	true	"User ID"
// @Success		200	{object}	store.User
// @Failure		400	{object}	error
// @Failure		404	{object}	error
// @Failure		500	{object}	error
// @Security		ApiKeyAuth
// @Router			/users/{id} [get]
func (app *application) getUserHandler(w http.ResponseWriter, r *http.Request) {

	user := getUserFromContext(r)

	if err := app.jsonResponse(w, http.StatusOK, user); err != nil {
		app.internalServerError(w, r, err)
		return
	}

}

type FollowUserPayload struct {
	UserID int64 `json:"user_id"`
}

// @Summary		Follow a user
// @Description	Follow a user by ID
// @Tags			users
// @Accept			json
// @Produce		json
// @Param			userID	path		int		true	"User ID"
// @Success		204		{string}	string	"User followed"
// @Failure		400		{object}	error	"USer payload missing or invalid"
// @Failure		404		{object}	error	"User not found"
// @Failure		500		{object}	error	"Server error"
// @Security		ApiKeyAuth
// @Router			/users/{userID}/follow [put]
func (app *application) followUserHandler(w http.ResponseWriter, r *http.Request) {
	followerUser := getUserFromContext(r)
	//TODO: Revert back to auth userID from ctx

	var payload FollowUserPayload
	if err := readJSON(w, r, &payload); err != nil {
		app.badRequestResponse(w, r, err)
		return
	}
	ctx := r.Context()
	// exists, err := app.store.Followers.ExistsFollow(ctx, followerUser.ID, payload.UserID)
	// if err != nil {
	// 	app.internalServerError(w, r, err)
	// 	return
	// }
	// if exists {
	// 	app.badRequestResponse(w, r, errors.New("user already followed"))
	// 	return
	// }
	if err := app.store.Followers.Follow(ctx, followerUser.ID, payload.UserID); err != nil {
		switch err {
		case store.ErrConflict:
			app.conflictResponse(w, r, err)
			return
		default:
			app.internalServerError(w, r, err)
			return

		}
	}

	if err := app.jsonResponse(w, http.StatusNoContent, nil); err != nil {
		app.internalServerError(w, r, err)
		return
	}

}

// UnfollowUser godoc
//
// @Summary		Unfollow a user
// @Description	Unfollow a user by ID
// @Tags			users
// @Accept			json
// @Produce		json
// @Param			userID	path		int		true	"User ID"
// @Success		204		{string}	string	"User unfollowed"
// @Failure		400		{object}	error	"USer payload missing or invalid"
// @Failure		404		{object}	error	"User not found"
// @Failure		500		{object}	error	"Server error"
// @Security		ApiKeyAuth
// @Router			/users/{userID}/unfollow [put]
func (app *application) unfollowUserHandler(w http.ResponseWriter, r *http.Request) {
	unfollowedUser := getUserFromContext(r)
	//TODO: Revert back to auth userID from ctx
	var payload FollowUserPayload
	if err := readJSON(w, r, &payload); err != nil {
		app.badRequestResponse(w, r, err)
		return
	}
	ctx := r.Context()

	if err := app.store.Followers.Unfollow(ctx, unfollowedUser.ID, payload.UserID); err != nil {
		app.internalServerError(w, r, err)
		return
	}
	if err := app.jsonResponse(w, http.StatusNoContent, nil); err != nil {
		app.internalServerError(w, r, err)
		return
	}

}
func (app *application) activateUserHandler(w http.ResponseWriter, r *http.Request) {
	token := chi.URLParam(r, "token")
	err := app.store.Users.Activate(r.Context(), token)
	if err != nil {
		switch err {
		case store.ErrRecordNotFound:
			app.notFoundResponse(w, r, err)
		default:
			app.internalServerError(w, r, err)
		}
		return
	}
	if err := app.jsonResponse(w, http.StatusNoContent, nil); err != nil {
		app.internalServerError(w, r, err)
		return
	}
}
func (app *application) userContextMiddleware(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		idParams := chi.URLParam(r, "userID")
		userID, err := strconv.ParseInt(idParams, 10, 64)
		if err != nil {
			app.badRequestResponse(w, r, err)
			return
		}
		ctx := r.Context()
		user, err := app.store.Users.GetByID(ctx, userID)
		if err != nil {
			switch {
			case errors.Is(err, store.ErrRecordNotFound):
				app.notFoundResponse(w, r, err)
			default:
				app.internalServerError(w, r, err)
			}
			return
		}
		ctx = context.WithValue(ctx, userCtx, user)

		next.ServeHTTP(w, r.WithContext(ctx))
	})
}
func getUserFromContext(r *http.Request) *store.User {
	user, _ := r.Context().Value(userCtx).(*store.User)
	return user
}